Alterei o botão de edição do usuario

+++ src/main/java/com/manhattan/demo/Dtos/ProductType/ProductTypeRequestDto.java
@@ -13,6 +13,6 @@ public record ProductTypeRequestDto(
         String tipo,
         @NotNull
         @PositiveOrZero
-        BigDecimal valor
+        float valor
 ) {
 }
